+++ app/src/main/java/com/github/felipecastilhos/pokedexandroid/features/pokemon/domain/usecase/PokemonUseCase.kt
@@ -1,7 +1,7 @@
 package com.github.felipecastilhos.pokedexandroid.features.pokemon.domain.usecase
 
-import com.github.felipecastilhos.pokedexandroid.features.pokemon.domain.repository.Pokemon
-import com.github.felipecastilhos.pokedexandroid.features.pokemon.domain.repository.PokemonMockRepository
+import com.github.felipecastilhos.pokedexandroid.features.pokemon.data.datasource.repository.Pokemon
+import com.github.felipecastilhos.pokedexandroid.features.pokemon.data.datasource.repository.PokemonMockRepository
 import javax.inject.Inject
 
 /**
