+++ app/src/main/java/com/github/felipecastilhos/pokedexandroid/features/pokemon/domain/viewmodel/PokemonListViewModel.kt
@@ -5,8 +5,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.github.felipecastilhos.pokedexandroid.core.coroutines.DispatcherProvider
 import com.github.felipecastilhos.pokedexandroid.core.logs.LogHandler
-import com.github.felipecastilhos.pokedexandroid.features.pokemon.domain.repository.Pokemon
-import com.github.felipecastilhos.pokedexandroid.features.pokemon.domain.repository.PokemonType
+import com.github.felipecastilhos.pokedexandroid.features.pokemon.data.datasource.repository.Pokemon
+import com.github.felipecastilhos.pokedexandroid.features.pokemon.data.datasource.repository.PokemonType
 import com.github.felipecastilhos.pokedexandroid.features.pokemon.domain.usecase.PokemonUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
